FEAT: determine graphiz output format from file extension
now if you use `--output-path dag.png`, you get a .png. This is useful to me
so I can embed the plot in my docs.

I considered adding a whole new `--format` CLI flag etc, but I
decided that was overkill. Could be added later if someone
really wants it.

Ironically, this is now consistent with _HELP_TEXT_OUTPUT