switch from uv run python to sys.executable
To ensure that the same python interpreter is used to spawn pytask as is
running the test we currently rely on `uv` to spawn the instance of
python. This relies on us having spawned the test process from `uv` in
the first place and requires that `uv` be used by anyone attempting to
test the software.

This switches the pattern `uv run python` to call
(`sys.executable`)[https://docs.python.org/3/library/sys.html#sys.executable]
which gives the absolute path of the interpreter being used.